+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.18.19
+### is.codion.swing.framework.ui
+- EntityTablePanel.Config.cellRenderer() and cellEditor() added.
 
 ## 0.18.18
 ### is.codion.common.db